Implemented argument parsing into a struct.
The line returned from readline() must be parsed and tokenized.
This has to be done twice, since argc needs to be known to
malloc argv in the next step.

The first go through the line is counting argc with a state
machine (if in a whitespace delimiter or out).

The second go through the line is with strsep() (like strtok())
and tokenizing on the delimiters (again), and building up argv.

argv then needs to be malloc'd, since it is on the stack.

A struct args holds the malloc'd argv, argc, the command,
and whether it runs in the foreground or the background.

Could decompose this function into smaller ones.
Also, can get rid of the argv on the stack, and only use the
malloc'd argv. Cleanup of debugging code is needed.