A few random things in this commit.
Make sure to set all unused pointers to NULL after the memory
they point to is free'd. args_free() should not set a = NULL
since that is a copy of the pointer. The caller has to explicitly
set their pointer to NULL after the call to *_free().

New function proc_free() which frees up a struct proc. The pointer
to the struct still needs to be set to NULL after memory is free'd.
proc_free() calls args_free(), so use proc_free() when dealing
with a struct proc.

Move the 'exit' command processing to proc_run() where it belongs.

More comments to clarify some tricky spots (mainly with the
'bg' command, which needs to shift all command arguments over).